rm SQL.Column SQL.Table
