+++ kore/src/SQL/Key.hs
@@ -15,12 +15,10 @@ import Prelude.Kore
 import Data.Int
     ( Int64
     )
-import Data.Proxy
 import qualified Generics.SOP as SOP
 import qualified GHC.Generics as GHC
 
 import Debug
-import SQL.Column
 
 {- | A foreign key into the table for type @a@.
  -}
@@ -36,7 +34,3 @@ instance SOP.HasDatatypeInfo (Key a)
 instance Debug (Key a)
 
 instance Diff (Key a)
-
-instance Column (Key a) where
-    defineColumn _ = defineColumn (Proxy @Int64)
-    toColumn = toColumn . getKey
